separating vineyard classes from plotting
