+++ src/python/doc/persistence_graphical_tools_sum.inc
@@ -2,13 +2,14 @@
    :widths: 30 40 30
 
    +-----------------------------------------------------------------+-----------------------------------------------------------------------+---------------------------------------------------------+
-   | .. figure::                                                     | These graphical tools comes on top of persistence results and allows  | :Authors: Vincent Rouvreau, Theo Lacombe                |
-   |      img/graphical_tools_representation.png                     | the user to display easily persistence barcode, diagram or density.   |                                                         |
-   |                                                                 |                                                                       | :Since: GUDHI 2.0.0                                     |
-   |                                                                 | Note that these functions return the matplotlib axis, allowing        |                                                         |
-   |                                                                 | for further modifications (title, aspect, etc.)                       | :License: MIT                                           |
+   | .. figure::                                                     | These graphical tools comes on top of persistence results and allows  | :Authors: Vincent Rouvreau, Theo Lacombe,               |
+   |      img/graphical_tools_representation.png                     | the user to display easily persistence barcode, diagram or density,   |           Hannah Schreiber                              |
+   |                                                                 | as well as vineyards and representative cycles.                       | :Since: GUDHI 2.0.0 and 3.1x.0                          |
    |                                                                 |                                                                       |                                                         |
+   |                                                                 | Note that these functions return the matplotlib axis, allowing        | :License: MIT                                           |
+   |                                                                 | for further modifications (title, aspect, etc.)                       |                                                         |
    |                                                                 |                                                                       | :Requires: `Matplotlib <installation.html#matplotlib>`_ |
    +-----------------------------------------------------------------+-----------------------------------------------------------------------+---------------------------------------------------------+
    | * :doc:`persistence_graphical_tools_user`                       | * :doc:`persistence_graphical_tools_ref`                                                                                        |
+   |                                                                 | * :doc:`vineyard_graphical_tools_ref`                                                                                           |
    +-----------------------------------------------------------------+---------------------------------------------------------------------------------------------------------------------------------+
