+++ src/python/doc/vineyard_graphical_tools_ref.rst
@@ -1,0 +1,10 @@
+:orphan:
+
+.. To get rid of WARNING: document isn't included in any toctree
+
+============================================
+Vineyard graphical tools reference manual
+============================================
+
+.. autofunction:: gudhi.vineyard_graphical_tools.plot_vineyards
+.. autofunction:: gudhi.vineyard_graphical_tools.plot_1D_representative_cycles
