+++ src/python/doc/vineyard_sum.inc
@@ -12,4 +12,5 @@
    |                                                                | The implementation is based on :cite:`vineyards` and :cite:`zigzag`.   |                                                                                    |
    +----------------------------------------------------------------+------------------------------------------------------------------------+------------------------------------------------------------------------------------+
    | * :doc:`vineyard_user`                                         | * :doc:`vineyard_ref`                                                                                                                                       |
+   |                                                                | * :doc:`vineyard_graphical_tools_ref`                                                                                                                       |
    +----------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------+
